Combined CharacterSheetCalculations.js, CircleToggle.js, edbitButtonHandlers.js and pageUtilis.js into one file: uiHandlers..js
