+++ index.html
@@ -1019,11 +1019,8 @@
     <script type="module" src="js/blockActionsHandler.js"></script>
     <script type="module" src="js/overlayHandler.js"></script>
     <script type="module" src="js/filterManager.js"></script>
-    <script type="module" src="js/circleToggle.js"></script>
-    <script type="module" src="js/editButtonHandler.js"></script>
+    <script type="module" src="js/uiHandlers.js"></script>
     <script type="module" src="js/diceRoller.js"></script>
-    <script type="module" src="js/CharacterSheetCalulations.js"></script>
-    <script src="js/pageUtils.js"></script>
 
 </body>
 </html>
